requirements.txt
astroid==2.2.5
coverage==4.4.2
Django==2.2.4
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
PyJWT==1.7.1